# [perf] Leverage spatial locality of reference for OTAP conversion