## Changes
- Append data to LogRecord batch in a column-wise fashion instead of row-wise fashion for better cache utilization
- `LogsRecordBatchBuilder` has a lot of fields which are array builders. The hypothesis is that when adding data for a large number of log records, it should be better to add data for multiple rows together since the builder's memory for that column should be hot in cache
- Since this is hot-path, we don't want to allocate a vec to hold all the log records. I'm chunking the data into 64 log records or less and storing them in an array

## Benchmarks
I have observed a ~5% increase in the benchmark performance.

The input for the existing benchmark does not have many log records within a scope. The performance benefit of this change is even more visible for cases (> 10%) when you have a high number of log records within a scope (for example, 1000, 5000).