@lquerel 2025-07-09
It looks cool and promising! What setup did you use to observe the results? proto_bytes -> views -> OTAP or a different configuration based on SYSLOG?

Could you also add the benchmark results before and after your optimization to the PR description?