@utpilla 2025-07-09
> It looks cool and promising! What setup did you use to observe the results? proto_bytes -> views -> OTAP or a different configuration based on SYSLOG?

I ran the existing benchmark for  proto_bytes -> views -> OTAP. Since this change is mainly concerned with the views -> OTAP conversion, I didn't really bother about the source of views as much.

> Could you also add the benchmark results before and after your optimization to the PR description?

For the benchmark: proto_bytes -> views -> OTAP

OS: Ubuntu 22.04.4 LTS (5.15.153.1-microsoft-standard-WSL2)
Hardware: Intel(R) Xeon(R) Platinum 8370C CPU @ 2.80GHz, 16vCPUs,
RAM: 64.0 GB

| Benchmark                             | main branch | PR branch | Perf Diff (%) |
|---------------------------------------|-------------|-----------|----------------|
| 5 Resources 10 Scopes 5 LogRecords (existing benchmark)    | 592.61 µs   | 566.23 µs |          4.45% |
| 1 Resource 1 Scope 1_000 LogRecords   | 2.1 ms      | 1.87 ms   |         10.95% |
| 1 Resource 1 Scope 5_000 LogRecords   | 10.1 ms     | 9.1 ms    |          9.90% |
| 1 Resource 1 Scope 10_000 LogRecords  | 21.08 ms    | 18.34 ms  |         13.00% |