@lquerel 2025-07-09
@utpilla Very interesting results! I was asking about the source because I was wondering whether the gains in data locality on one side (views -> OTAP) could potentially be offset on the other side (proto_bytes -> views). But it seems the balance is favorable, so that’s great.

I’ll approve the PR, but I suggest leaving it open until @albertlockett is back so he can see those results as well.

@msalib, could you check on your side if you see the same effects for traces?